:add: paideia strategic contribute
